Refacto auth middle so the tracer is not passed to the handler
