+++ package.json
@@ -29,7 +29,7 @@
   "dependencies": {
     "@aws-sdk/client-dynamodb": "3.4.1",
     "@aws-sdk/util-dynamodb": "3.4.1",
-    "@fewlines/connect-client": "0.2.5",
+    "@fewlines/connect-client": "0.4.0",
     "@fewlines/connect-management": "0.2.1",
     "@fwl/logging": "0.1.1",
     "@fwl/tracing": "0.9.0",
