+++ src/pages/account/logins/[type]/validation/[eventId].tsx
@@ -15,7 +15,7 @@ import { Container } from "@src/components/containers/container";
 import { ValidateIdentityForm } from "@src/components/forms/validate-identity-form";
 import { Layout } from "@src/components/page-layout";
 import { logger } from "@src/logger";
-import { withAuth } from "@src/middlewares/with-auth";
+import { authMiddleware } from "@src/middlewares/auth-middleware";
 import { withSentry } from "@src/middlewares/with-sentry";
 import getTracer from "@src/tracer";
 
@@ -42,18 +42,16 @@ const ValidateIdentityPage: React.FC<{
   );
 };
 
-const tracer = getTracer();
-
 const getServerSideProps: GetServerSideProps = async (context) => {
   return getServerSidePropsWithMiddlewares<{ type: string }>(
     context,
     [
-      tracingMiddleware(tracer),
-      recoveryMiddleware(tracer),
+      tracingMiddleware(getTracer()),
+      recoveryMiddleware(getTracer()),
       withSentry,
-      errorMiddleware(tracer),
-      loggingMiddleware(tracer, logger),
-      withAuth,
+      errorMiddleware(getTracer()),
+      loggingMiddleware(getTracer(), logger),
+      authMiddleware(getTracer()),
     ],
     "/account/logins/[type]/validation/[eventId]",
     async (request, response: ServerResponse) => {
